pyproject.toml
[tool.poetry]
name = "simplec"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
attrs = "^21.4.0"
click = "^8.0.4"
antlr4-python3-runtime = "4.7.2"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
scc = 'simplec.cli:compiler'
